GH-45892: [Ruby] Unify test for list array in raw_records and each_raw_record
### Rationale for this change

The PR reduces duplicated test cases and ensures that both `raw_records` and `each_raw_record` behave consistently by extracting their common test cases.
- `Arrow::RecordBatch#each_raw_record`
- `Arrow::Table#each_raw_record`
- `Arrow::RecordBatch#raw_records`
- `Arrow::Table#raw_records`

### What changes are included in this PR?

We extracted shared test cases about list array used by both raw_records and each_raw_record.

### Are these changes tested?

Yes.

### Are there any user-facing changes?

No.

* GitHub Issue: close GH-45892
